Import v1.6.6 r1
Source: http://www.openfst.org/twiki/bin/viewfile/FST/FstDownload?filename=openfst-1.6.6.tar.gz;rev=1

+++ src/include/fst/prune.h
@@ -194,7 +194,7 @@ void Prune(MutableFst<Arc> *fst, const PruneOptions<Arc, ArcFilter> &) {
 template <class Arc>
 void Prune(MutableFst<Arc> *fst, typename Arc::Weight weight_threshold,
            typename Arc::StateId state_threshold = kNoStateId,
-           double delta = kDelta) {
+           float delta = kDelta) {
   const PruneOptions<Arc, AnyArcFilter<Arc>> opts(
       weight_threshold, state_threshold, AnyArcFilter<Arc>(), nullptr, delta);
   Prune(fst, opts);
